Refactor: Split utils into separate files
1. Extract each function to a separate file.
2. Add globalizeComponents function, so getRouteData is now pure.
3. Replace Object.assign with spread operator.
4. Use deepfreeze in tests to prevent mutations.
