+++ package.json
@@ -101,6 +101,7 @@
     "babel-preset-react": "^6.24.1",
     "cheerio": "^1.0.0-rc.2",
     "danger": "^2.0.1",
+    "deepfreeze": "^2.0.0",
     "dog-names": "^1.0.2",
     "enzyme": "^3.1.1",
     "enzyme-adapter-react-16": "^1.0.4",
