+++ src/rsg-components/slots/IsolateButton.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import MdFullscreen from 'react-icons/lib/md/fullscreen';
 import MdFullscreenExit from 'react-icons/lib/md/fullscreen-exit';
 import ToolbarButton from 'rsg-components/ToolbarButton';
-import { getUrl } from '../../utils/utils';
+import getUrl from '../../utils/getUrl';
 
 const IsolateButton = ({ name, example, isolated }) =>
 	isolated ? (
