+++ src/utils/__tests__/filterComponentsByExactName.spec.js
@@ -1,0 +1,18 @@
+import deepfreeze from 'deepfreeze';
+import filterComponentsByExactName from '../filterComponentsByExactName';
+
+const components = deepfreeze([
+	{
+		name: 'Button',
+	},
+	{
+		name: 'Image',
+	},
+]);
+
+describe('filterComponentsByExactName', () => {
+	it('should return components with exact name', () => {
+		const result = filterComponentsByExactName(components, 'Image');
+		expect(result.map(x => x.name)).toEqual(['Image']);
+	});
+});
