+++ src/utils/__tests__/globalizeComponent.spec.js
@@ -1,0 +1,14 @@
+import globalizeComponent from '../globalizeComponent';
+
+describe('globalizeComponent', () => {
+	it('should set component’s module as a global variable', () => {
+		const globalsCount = Object.keys(global).length;
+		globalizeComponent({
+			name: 'Foo',
+			props: {},
+			module: 13,
+		});
+		expect(Object.keys(global).length).toBe(globalsCount + 1);
+		expect(global.Foo).toBe(13);
+	});
+});
