+++ src/utils/filterComponentsInSectionsByExactName.js
@@ -1,0 +1,21 @@
+import filterComponentsByExactName from './filterComponentsByExactName';
+
+/**
+ * Recursively filters all components in all sections by component name.
+ *
+ * @param {object} sections
+ * @param {string} name
+ * @return {Array}
+ */
+export default function filterComponentsInSectionsByExactName(sections, name) {
+	const components = [];
+	sections.forEach(section => {
+		if (section.components) {
+			components.push(...filterComponentsByExactName(section.components, name));
+		}
+		if (section.sections) {
+			components.push(...filterComponentsInSectionsByExactName(section.sections, name));
+		}
+	});
+	return components;
+}
