+++ src/utils/filterSectionExamples.js
@@ -1,0 +1,14 @@
+/**
+ * Return a copy of the given section with the examples array filtered
+ * to contain only the specified index
+ *
+ * @param {object} section
+ * @param {number} index
+ * @returns {object}
+ */
+export default function filterSectionExamples(section, index) {
+	return {
+		...section,
+		content: [section.content[index]],
+	};
+}
