+++ src/utils/processSections.js
@@ -1,0 +1,15 @@
+import processComponents from './processComponents';
+
+/**
+ * Recursively process each component in all sections.
+ *
+ * @param {Array} sections
+ * @return {Array}
+ */
+export default function processSections(sections) {
+	return sections.map(section => ({
+		...section,
+		components: processComponents(section.components || []),
+		sections: processSections(section.sections || []),
+	}));
+}
